requirements.txt
streamlit
gitpython
nbformat
langchain>=0.1.16
langchain-core>=0.1.52
langchain-community>=0.1.52
langchain-groq
sentence-transformers
faiss-cpu
